View, create and delete api keys via UI
- Add discord user id as internal username to ensure uniqueness
- Use Integer for internal user id to match database schema
- Remove unused Methods from UserPrincipal

+++ src/main/java/de/sustineo/simdesk/entities/auth/UserPermission.java
@@ -6,7 +6,7 @@ import java.time.Instant;
 
 @Data
 public class UserPermission {
-    private Long userId;
+    private Integer userId;
     private UserRoleEnum role;
     private Instant insertDatetime;
 }
